libc: Re-enable remaining warnings
Fix integer sign comparison warnings by using suitable types.
Also adapt coding style a bit where suitable.

+++ src/lib/libc/src/internal/intscan.c
@@ -69,11 +69,11 @@ unsigned long long __intscan(FILE *f, unsigned base, int pok, unsigned long long
 		}
 	}
 	if (base == 10) {
-		for (x=0; c-'0'<10U && x<=UINT_MAX/10-1; c=shgetc(f))
+		for (x=0; (unsigned)(c - '0') < 10U && x<=UINT_MAX/10-1; c=shgetc(f))
 			x = x*10 + (c-'0');
-		for (y=x; c-'0'<10U && y<=ULLONG_MAX/10 && 10*y<=ULLONG_MAX-(c-'0'); c=shgetc(f))
+		for (y=x; (unsigned)(c - '0') < 10U && y<=ULLONG_MAX/10 && 10*y<=ULLONG_MAX-(c-'0'); c=shgetc(f))
 			y = y*10 + (c-'0');
-		if (c-'0'>=10U) goto done;
+		if ((unsigned)(c - '0') >= 10U) goto done;
 	} else if (!(base & (base - 1))) {
 		int bs = "\0\1\2\4\7\3\6\5"[(0x17*base)>>5&7];
 		for (x=0; val[c]<base && x<=UINT_MAX/32; c=shgetc(f))
